Add the debian directory.
This unintrusive directory contains all the necessary code
to build the Debian package. Anybody on a Debian/Ubuntu system
can leverage that directory contents to build a package.

This directory builds upon the CMake-based configuration.
However, they are totally dissociated.

Having this directory in the source tree makes it easy to
create a new Debian package by simply issuing the command
'dpkg-buildpackage'.