Remove patches that installed the modifications to the CMake-based build system.
These patches used to install the large modifications to the CMake-based build
system. Since these modifications are in the process of being integrated
upstream, they are no more useful. We remove them now.

We only keep the patches that modify upstram code so that it still works when
the Debian package is installed (header file include directives, for example).