Change CMake install modules dir to cmake/isospec++ (++ were missing)
And many other minor fixes/improvements/additions:

Implementing the handling of the documentation.

Change the install directory of the doc because the doc
if for the whole package and not only for the C++ lib.

Fix the bad substitutions in the pkg-config file, so as to preserve
strings like ${prefix}. Use the @ONLY argument to the configure_file
function call.